+++ packages/react-query/src/useSuspenseInfiniteQuery.ts
@@ -7,8 +7,9 @@ import {
   parseQueryArgs,
   useInfiniteQuery,
 } from '@tanstack/react-query'
+import type { OmitKeyof } from './utility-types'
 
-export type BaseUseSuspenseInfiniteQueryResult<TData = unknown> = Omit<
+export type BaseUseSuspenseInfiniteQueryResult<TData = unknown> = OmitKeyof<
   UseInfiniteQueryResult<TData>,
   'error' | 'isLoadingError' | 'isError' | 'isRefetchError' | 'isFetching'
 > & { status: 'success' | 'loading' }
@@ -31,7 +32,7 @@ export type UseSuspenseInfiniteQueryOptions<
   TError = unknown,
   TData = TQueryFnData,
   TQueryKey extends QueryKey = QueryKey,
-> = Omit<UseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryFnData, TQueryKey>, 'suspense'>
+> = OmitKeyof<UseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryFnData, TQueryKey>, 'suspense'>
 
 /**
  * This hook wrapping useInfiniteQuery of @tanstack/react-query with default suspense option. with this hook, you don't have to make unnecessary type narrowing
@@ -46,7 +47,7 @@ export function useSuspenseInfiniteQuery<
 >(
   queryKey: TQueryKey,
   queryFn: QueryFunction<TQueryFnData, TQueryKey>,
-  options?: Omit<
+  options?: OmitKeyof<
     UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>,
     'enabled' | 'queryKey' | 'queryFn'
   >
@@ -59,7 +60,7 @@ export function useSuspenseInfiniteQuery<
 >(
   queryKey: TQueryKey,
   queryFn: QueryFunction<TQueryFnData, TQueryKey>,
-  options: Omit<
+  options: OmitKeyof<
     UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>,
     'enabled' | 'queryKey' | 'queryFn'
   > & {
@@ -74,7 +75,7 @@ export function useSuspenseInfiniteQuery<
 >(
   queryKey: TQueryKey,
   queryFn: QueryFunction<TQueryFnData, TQueryKey>,
-  options: Omit<
+  options: OmitKeyof<
     UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>,
     'enabled' | 'queryKey' | 'queryFn'
   > & {
@@ -89,7 +90,7 @@ export function useSuspenseInfiniteQuery<
 >(
   queryKey: TQueryKey,
   queryFn: QueryFunction<TQueryFnData, TQueryKey>,
-  options: Omit<UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'queryKey' | 'queryFn'>
+  options: OmitKeyof<UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'queryKey' | 'queryFn'>
 ): UseSuspenseInfiniteQueryResultOnSuccess<TData> | UseSuspenseInfiniteQueryResultOnLoading
 
 // arg1: queryKey, arg2: options
@@ -100,7 +101,7 @@ export function useSuspenseInfiniteQuery<
   TQueryKey extends QueryKey = QueryKey,
 >(
   queryKey: TQueryKey,
-  options?: Omit<UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'enabled' | 'queryKey'>
+  options?: OmitKeyof<UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'enabled' | 'queryKey'>
 ): UseSuspenseInfiniteQueryResultOnSuccess<TData>
 export function useSuspenseInfiniteQuery<
   TQueryFnData = unknown,
@@ -109,7 +110,10 @@ export function useSuspenseInfiniteQuery<
   TQueryKey extends QueryKey = QueryKey,
 >(
   queryKey: TQueryKey,
-  options: Omit<UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'enabled' | 'queryKey'> & {
+  options: OmitKeyof<
+    UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>,
+    'enabled' | 'queryKey'
+  > & {
     enabled?: true
   }
 ): UseSuspenseInfiniteQueryResultOnSuccess<TData>
@@ -120,7 +124,10 @@ export function useSuspenseInfiniteQuery<
   TQueryKey extends QueryKey = QueryKey,
 >(
   queryKey: TQueryKey,
-  options: Omit<UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'enabled' | 'queryKey'> & {
+  options: OmitKeyof<
+    UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>,
+    'enabled' | 'queryKey'
+  > & {
     enabled: false
   }
 ): UseSuspenseInfiniteQueryResultOnLoading
@@ -131,7 +138,7 @@ export function useSuspenseInfiniteQuery<
   TQueryKey extends QueryKey = QueryKey,
 >(
   queryKey: TQueryKey,
-  options: Omit<UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'queryKey'>
+  options: OmitKeyof<UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'queryKey'>
 ): UseSuspenseInfiniteQueryResultOnSuccess<TData> | UseSuspenseInfiniteQueryResultOnLoading
 
 // arg1: options
@@ -141,7 +148,7 @@ export function useSuspenseInfiniteQuery<
   TData = TQueryFnData,
   TQueryKey extends QueryKey = QueryKey,
 >(
-  options: Omit<UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'enabled'> & {
+  options: OmitKeyof<UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'enabled'> & {
     enabled?: true
   }
 ): UseSuspenseInfiniteQueryResultOnSuccess<TData>
@@ -151,7 +158,7 @@ export function useSuspenseInfiniteQuery<
   TData = TQueryFnData,
   TQueryKey extends QueryKey = QueryKey,
 >(
-  options: Omit<UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'enabled'> & {
+  options: OmitKeyof<UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'enabled'> & {
     enabled: false
   }
 ): UseSuspenseInfiniteQueryResultOnLoading
@@ -174,8 +181,8 @@ export function useSuspenseInfiniteQuery<
   arg1: TQueryKey | UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>,
   arg2?:
     | QueryFunction<TQueryFnData, TQueryKey>
-    | Omit<UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'queryKey'>,
-  arg3?: Omit<UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'queryKey' | 'queryFn'>
+    | OmitKeyof<UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'queryKey'>,
+  arg3?: OmitKeyof<UseSuspenseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryKey>, 'queryKey' | 'queryFn'>
 ) {
   return useInfiniteQuery({
     ...parseQueryArgs(arg1, arg2, arg3),